agent: Wait to regenerate restore endpoints until ipcache has been populated
[ upstream commit 2ee4d5f2c323df859ec53a093fe116ec0a948c27 ]

In the process of endpoints being restored from disk, all programs are
regenerated. This bases the programs on the latest snapshots of all local
caches for identities, ipcache, services, and so on. In order to not
destabilize existing endpoint, it must be ensured that local caches have been
synchronized to a valid state. This was done for services as well as identities
if the endpoint was not assigned a fixed or well-known identity. The same
logic is required for the ipcache to avoid egress drops.

Signed-off-by: Thomas Graf <thomas@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>