endpoints: Add optional callback to WaitForPolicyRevision
[ upstream commit ce456f63665b33e6a7003ca80aafc8db06a2a00b ]

Endpoints already have a notification scheme for when they reach certain
policy revisions. This is convenient for setting true/false values like
the Enforcing field on CNPs. It is more difficult to measure when the
revision is reached by the endpoint because the goroutine waiting on the
channel returned by WaitForPolicyRevision may not be scheduled
immediately. Using a callback means arbitrary code can be run, ideally
non-blocking. We also pass in the timestamp to ensure a more accurate
time measurement is possible.

Signed-off-by: Ray Bejjani <ray@covalent.io>
Signed-off-by: Joe Stringer <joe@cilium.io>